reworked IDs in various systems, polishing stats system

+++ Source/VUNDK/Public/Features/Gameplay/RPGSystem/StatsSystem/Containers/SpecialStatsContainer.h
@@ -18,5 +18,5 @@ public:
 	void AddSpecialStats(TSet<USpecialStatData*> SpecialStats);
 
 	UFUNCTION(BlueprintPure)
-	TMap<USpecialStatData*, int32> GetSpecialStatsValues();
+	TMap<USpecialStatData*, float> GetSpecialStatsValues();
 };
